Corrected YCbCr -> RGB conversion
The code I snarfed from Cortado was borked in some way and was
not quite the right color. Redid using floating point from
formula in Wikipedia. ;)

Not sure if it's best to use floating point here, but default
JS engines would be doing math as FP anyway. Should fix the int
version and benchmark some time.