Optionally rebuild the swf flash shim (needs Apache Flex SDK)
Still including a build artifact in the source tree so you
don't have to go installing Apache Flex SDK if you're not
going to change the Flash audio shim, but it's not too hard.

Make sure the Flex SDK 'bin' dir (with 'mxmlc' etc) is in the
PATH when running 'make swf'.