Halve the time spent in TypedArray.set() in IE 11
Making the hopefully not unwarranted assumption that we can copy
as dwords instead of bytes from the ycbcr array.

IE's TypedArray.set() is horribly slow and seems to basically do
a for loop and set each element individually. Copying the ycbcr
frame buffers out 32 bits at a time seems to save noticeably.