Switch from tremor back to libvorbis for performance
It turns out that tremor is actually significantly slower in the
JS virtual machine on slower devices like an iPod Touch 5th-generation.
A *lot* of time gets spent in the integer multiplication functions.

Going back to vorbis shaved a ~60ms per-frame decode time down to ~40ms
or less on the iPod on the 160p/15fps videos.