Greatly speed up WebGL on IE 11
IE 11 lacks a way to upload luminance-only or alpha-only textures;
you can only upload them as RGB or RGBA.

This was requiring a slow conversion step on the Y, Cb, and Cr byte
planes to expand them into int planes. This pre-upload conversion in
JS has been replaced with render-time conversion on the GPU...

The byte textures are uploaded as pseudo-RGBA textures at 1/4 the width
and then their components are unpacked into separate texels at sample
time, using a static 'stripe' texture that's prefilled with a multiplier
to select the right channel for each texel.

Amazingly this works pretty well on IE 11 and doesn't seem to slow down
on other browsers that accept luminance textures, so skipping logic
to select the mode and doing it unconditionally for now.

https://github.com/brion/ogv.js/issues/94