Placebo workaround for iPad 3 possible JIT bug
There's some kind of problem with the JIT in iOS 7 Safari
that sometimes trips up on optimized Vorbis builds, at least
on my iPad 3 (A5X SoC).

Exercising some of the ogg & vorbis library code paths with
a second decoder for the first few packets of data seems to
be enough to work around this.

Non-deterministic debugging ROCKS!