Fix perf regression in IE+Flash audio (buffer consolidation)
Reimplemented a second buffer stage between audio decode and
sending hex data to Flash; this cuts overhead of buffering the
audio data significantly on slower machines where it was as
expensive as decoding the audio itself.

And did it much simplerer than the old method that was in the
middle of the decode thingy. :D