Bypass audioFeeder when starting playback muted
This switches playback to drive off of the video clock timers
when started muted, which allows script-triggered autoplay on
iOS of files containing audio.

May or may not have surprise bugs in the muted path.

Audio is still decoded, so when unmuting the audioFeeder is
started up and we're able to continue from the current position.

Partial fix for https://github.com/brion/ogv.js/issues/112

Todo:
* try to autodetect the case where audio initialization failed
* clean up remaining bugs in non-audio clock