WebM demuxer fixes using updated nestegg
nestegg can now recover parsing context from running out of data
partway through a packet; this lets us avoid the need to keep a
bunch of extra data for multiple packets buffered, and lets us
run through to the end of the file ok.

Replaced the primitive buffer with a slightly fancier buffer
that is also easier to shift through.