Implement seeking for WebM with cues
Added callback allowing demuxer to request a seek on the input
stream. Along with new seekToKeypoint method, this lets us use
nestegg's synchronous seeking method by restarting it after seeks.
This is kind of reliant on the internal implementation, as we
just have to know that it's safe to restart. :P

When seeking the first time, we may need to wait for an initial
seek to the actual location of the cues in the file, as they tend
to be at end instead of at beginning.

Then we have to make sure that the input buffer holds the entire
cues element before reading it, or things will get confused.
This is done by reading the EBML element id and size, then seeking
back within the input buffer.

Then, we can call back into the seek, and it may trigger the final
seek on the input stream... at which point the linear seek to target
over in OGVPlayer takes over reading packets from the keyframe to
the actual target frame.

Yay! Seems to work and stuff.