skip failIfMajorPerformanceCaveat
it's too aggressive and was disabling webgl on devices
like a 2011 MacBook Pro running Windows 7, with perfectly
good performance.

Recommend in future if it's needed to blacklist things,
just run live and detect if we're absurdly slow then switch
back to software after a few frames. would require making a
new canvas possibly.