Stop audio to fix sync even if a frame isn't ready
This helps with cases where the decoder suddenly becomes
very slow, as on slow devices when first few seconds don't
trigger a fully JIT and then more detail comes in.

(Example: Tears of Steel 240p on Surface RT; when the
spaceship engines fire, the JIT fires up and it takes
about two seconds to get back on track. Previously, audio
continued running until the expensive frame came back,
then a bunch of frames ran in quick succession. Now, the
audio stops right away and when it continues, we're running
smooth the whole time.)