1.1.3 - video.js seek bug fix
fix play-during-seek bug that interacted with video.js badly

Video.js's seek bar was firing lots of pause and play to keep
things paused during scrubbing, then continue when done. This
caused a breakage when clicking quickly, as we tried to resume
playback while the seek was still going and it didn't _quite_
like that. :)

Didn't turn up in own test harness, which left play on during
the scrubbing.