First stage use of new stream-file module
Switched to the new promises-based StreamFile interface on
OGVPlayer, using the existing 'push' model for demuxers.

Next step is to redo the demuxers to accept the StreamFile
object and manage their i/o themselves. This will move the
logic for Ogg bisection streaming over to the Ogg demuxer,
and will remove the need for the 'onseek' callback.

Todo:
* retool demuxer to take the stream
* fix options for cache size