Update audio-feeder to 0.4.6 for Flash coalescing
Small audio packets are now flushed to the Flash audio path
on IE 11 in larger chunks (40ms timeout). This reduces the number
of times we have to communicate with the Flash plugin, which is
expensive in IE on slow machines. Frees up significant time
on main thread on Dell Inspiron Duo Atom N570 1.6GHz