Fix onstarved handling
With the new sync code at least, was seeing a lot of cases on the
slow Atom machine where audio got starved and this ended up
'hanging', with the video never proceeding because audio got hung
inserting extra spaces and never catching up.

Now, when starved, switch back to prebuffering mode and halt audio
until it's ready to proceed -- this seems to work much better.
Also updated audio-feeder with a fix for this pattern so it doesn't
add extra bits and get us out of sync when we try to restart.

Fixes https://github.com/brion/ogv.js/issues/466