Allow linear seeking of WebM files without cues
This works well for audio files which are usually relatively small
and often already in buffer, but could be really slow for video.
Previously, such videos would just fail on seek attempt, but in
practice WebM videos usually come with cues.

The existing linear-seek from the keypoint to the target is used
in the high-level player; the low-level webm demuxer is changed to
reset to the beginning of the file if no cues are present.

Fixes https://github.com/brion/ogv.js/issues/489