Fix for audio/webm linear seeking
The initial position was incorrect because nestegg had already
read the start of the cluster -- this caused the first few packets
after seek to be emitted with wrong timestamps from nestegg,
causing the high-level linear seek to assume it was already at
target and thus play back the beginning of the file instead of
the desired position.

For now just magic-numbering the number of bytes it read ahead.
Replace this with a cleaner solution.

Fixes https://github.com/brion/ogv.js/issues/489