Micro-optimize copying of video buffers on IE 11
Avoid creating a separate TypedArray instance for each
line to copy out of the buffer. This is moderately faster
on IE 11 than using TypedArray.set() as well, and shows
a several percent throughput improvement on 480p Theora.
At 240p the difference is smaller, and gets drowned out
by decoder time on VP9.