Handle WebM discard-padding on audio packets
Fixes https://github.com/brion/ogv.js/issues/509

This doesn't resolve _all_ possible issues with funky
PTS on audio packets, but does fix it up if you're using
discard-padding to make packets variable-duration as in
the sample file on that task.

Previously the padding was played to audio and counted
against synchronized time, leaving the video to play
"early" and out of sync.

Now trims the audio packets as specified, getting the
sample file playing in sync.