Multithreaded version of av1 decoder
This went pretty smooth; rebuilt dav1d with pthreads on,
and refactored the threading support from the vpx codec
into a common helper header.

Not sure I'm correctly handling all cases of the encoder
as it implies that packets might not always produce output
and multiple outputs are possible, and that's maybe not
gonna work out correctly in the current code. But it seems
to work so far so :D

Only seeing modest improvement in decode performance on the
2-core/4-thread MacBookPro in Firefox, but it's visible.
The test files I have might also not be tiled correctly,
so will test again on new files soon.