Fix for replay-after-end-of-playback in WebM
In WebM streams, after playback we sometimes end up left exactly
at the beginning of the cues in netegg's input buffer.

When restarting playback, it would try to seek back to time 0,
which causes the netegg to seek to the beginning of cues and
load them so it knows where to seek to for realsies.

This ended up failing because we had flushed the demuxer data,
which in bq_flush() failed to reset the position -- causing
the seek to think it was already at the right position, but we
had no input data. As a result, no seek on the inputStream end
got issued, and no more data came in!

Fix is to reset the position to 0 on flush. Might be even more
robust to use a -1 invalid sigil but meh.

Fixes https://github.com/brion/ogv.js/issues/495