adjustment for mono stereo conversion
WebAudio, at least in Firefox and Chrome, seems to handle single-
channel output by duplicating to the left and right channels which
is louder than playing the audio by itself (at least on macOS).

Flash path did the same in our resampling end, since it's required
to output stereo there.

Now forcing WebAudio to stereo as well, and doing stereo upconversion
from mono with an adjustment factor of 1 / sqrt(2) which makes it
sound equivalent loudness to a single speaker, in theory.

This roughly matches the loudness of playing a mono file directly
in Firefox or Chrome.