Fix for strict mode in IE/Flash path; 0.4.20 tag
This came up in the videojs-ogvjs test running on a Windows machine
in IE; the way it was embedding ogv.js code ended up wrapping into
strict mode, which failed in this particular old bit of inherited
code that used `arguments.callee`.

Wasn't really needed (it was a func-call-to-new-operator bounce for
code used only internally, so always with 'new') so easy to just
remove.