Enable multithreaded SIMD build
MT gives a big boost over single-threaded, and works now
with LLVM upstream backend, so let's combine with SIMD.

Reaches 1080p24 reliably on even the tough parts of the
Camindandes 3 test video on my 2018 MacBook Pro 15" i9
(6-core).

Might not be worth doing separate MT and non-MT builds
for SIMD, depending on when anything comes out with it...
Consider just doing an MT build and dropping the single-
threaded one.