Copy vpx frames on decode thread for wasm-mt
Speeds things up a little on threaded builds since we don't
have to wait for the main thread to be free to go on to the
next queued frame decode.

Currently this still does a copy on the main thread later,
though it seems to be faster to do so somehow. o_O

Should probably try to avoid the final copy later, with a
callback to free or something.