Recover most of the lost performance in IE 11
Copying frame buffers with set on a Uint8Array or Uint32Array
is crazy slow on IE 10/11. On 11, slice is faster but that
creates a new buffer which must be freed, and we're trying
to recycle buffers here.

Using a Float64Array copies significantly faster, and comes
closer to the previous overhead of the copy. Whether the
reduction in buffers to GC makes up for it is unknown. :)