Fix build for IE
Needed to specify options to use an old browser version to get
the old emscripten_memcpy_big working but needed to _not_ specify
IE itself or it includes the Promise polyfill that breaks Closure.
We need Closure to keep the wasm2js output small.

Promise polyfill is provided separately in the ogv.js build
by the loaders in main thread and workers, so this is fine. :)