Bump to 1.8.8
Fix regression in Firefox/macOS with update to yuv-canvas 1.2.10.
Haven't figured out the cause of it, whether it's on our end or
in Firefox or in macOS's OpenGL implementation.

Worked around by disabling the affected unpacking mode in yuv-canvas
and using luminance textures always. This is simpler but sometimes
is slower depending on the implementation.

On macOS 12.3 this causes a noticeable reduction in throughput
but it's somewhat compensated for by improvements to recycling
texture buffers. In any case it's enough for target resolutions
but the higher you go it becomes more noticeable, as drawing time
blocking the main thread.