WIP retooling buffer extraction
Minimize buffer extraction to pack the minimal stride.
This is necessry to be compatible with the VideoFrame constructor
on Safari when using WebCodecs for *YUV blitting* but not for
decoding (eg when using a codec not supported by the browser,
but WebCodecs is present)

Chrome handled this fine by using our specified layout but
Safari seems to have trouble.